增加玩家死亡复活功能 增加物体属性配置 重构部分代码

+++ Assets/Scripts/GameEntity/Data/HurtData.cs
@@ -9,7 +9,7 @@ namespace Runtime
 
         public HurtData(string sender, string receiver) : base("HurtData", sender, receiver) { }
         public Vector2 AttackForce { get; set; }
-        public int AttackValue { get; set; }
+        public float AttackValue { get; set; }
         public int AttackerID {get;set;}
         public bool IsSwoon { get; set; }//是否击飞
         public float AttackerDir { get; set; }
